refactor: create external certificate own dir
